chore(table): new table examples added to storybook
